extract fetchRolesTree to utils
