+++ api/src/services/comments.ts
@@ -82,7 +82,7 @@ export class CommentsService extends ItemsService {
 				role: user['role']?.id ?? null,
 				admin: false,
 				app: false,
-				roles: await fetchRolesTree(user['role']?.id ?? null, this.knex),
+				roles: await fetchRolesTree(user['role']?.id ?? null, { knex: this.knex }),
 				ip: null,
 			};
