+++ api/src/services/roles.ts
@@ -4,8 +4,8 @@ import { UserIntegrityCheckFlag } from '@directus/types';
 import { clearSystemCache } from '../cache.js';
 import { fetchRolesTree } from '../permissions/lib/fetch-roles-tree.js';
 import { transaction } from '../utils/transaction.js';
-import { ItemsService } from './items.js';
 import { AccessService } from './access.js';
+import { ItemsService } from './items.js';
 import { PresetsService } from './presets.js';
 import { UsersService } from './users.js';
 
@@ -110,7 +110,7 @@ export class RolesService extends ItemsService {
 			throw new InvalidPayloadError({ reason: 'A role cannot be a parent of itself' });
 		}
 
-		const roles = await fetchRolesTree(parent, this.knex);
+		const roles = await fetchRolesTree(parent, { knex: this.knex });
 
 		if (ids.some((id) => roles.includes(id))) {
 			// The role tree up from the parent already includes this role, so it would create a circular reference
