+++ api/src/utils/get-accountability-for-token.ts
@@ -34,7 +34,7 @@ export async function getAccountabilityForToken(
 			if (payload.id) accountability.user = payload.id;
 
 			accountability.role = payload.role;
-			accountability.roles = await fetchRolesTree(payload.role, database);
+			accountability.roles = await fetchRolesTree(payload.role, { knex: database });
 
 			const { admin, app } = await fetchGlobalAccess(accountability, database);
 
@@ -56,7 +56,7 @@ export async function getAccountabilityForToken(
 
 			accountability.user = user.id;
 			accountability.role = user.role;
-			accountability.roles = await fetchRolesTree(user.role, database);
+			accountability.roles = await fetchRolesTree(user.role, { knex: database });
 
 			const { admin, app } = await fetchGlobalAccess(accountability, database);
