+++ packages/utils/node/fetch-roles-tree.test.ts
@@ -1,6 +1,6 @@
 import type { Knex } from 'knex';
 import { beforeEach, expect, test, vi } from 'vitest';
-import { _fetchRolesTree } from './fetch-roles-tree.js';
+import { fetchRolesTree } from './fetch-roles-tree.js';
 
 let knex: Knex;
 
@@ -16,7 +16,7 @@ beforeEach(() => {
 });
 
 test('Returns empty array if start value is null', async () => {
-	const roles = await _fetchRolesTree(null, knex);
+	const roles = await fetchRolesTree(null, { knex });
 	expect(roles).toEqual([]);
 });
 
@@ -26,7 +26,7 @@ test('Returns array of all parents in top-down order', async () => {
 	vi.mocked(knex.first).mockResolvedValueOnce({ id: 'third', parent: null });
 	vi.mocked(knex.first).mockResolvedValueOnce({ id: 'unrelated', parent: null });
 
-	const roles = await _fetchRolesTree('start', knex);
+	const roles = await fetchRolesTree('start', { knex });
 
 	expect(roles).toEqual(['third', 'second', 'start']);
 });
@@ -37,7 +37,7 @@ test('Exits if parent row is undefined', async () => {
 	vi.mocked(knex.first).mockResolvedValueOnce(undefined);
 	vi.mocked(knex.first).mockResolvedValueOnce({ id: 'unrelated', parent: null });
 
-	const roles = await _fetchRolesTree('start', knex);
+	const roles = await fetchRolesTree('start', { knex });
 
 	expect(roles).toEqual(['second', 'start']);
 });
@@ -47,7 +47,7 @@ test('Throws error if infinite recursion occurs', async () => {
 	vi.mocked(knex.first).mockResolvedValueOnce({ id: 'second', parent: 'third' });
 	vi.mocked(knex.first).mockResolvedValueOnce({ id: 'third', parent: 'first' });
 
-	await expect(_fetchRolesTree('first', knex)).rejects.toMatchInlineSnapshot(
+	await expect(fetchRolesTree('first', { knex })).rejects.toMatchInlineSnapshot(
 		`[Error: Recursion encountered: role "third" already exists in tree path "third"->"second"->"first"]`,
 	);
 });
