+++ packages/utils/node/index.ts
@@ -1,4 +1,5 @@
 export * from './array-helpers.js';
+export * from './fetch-roles-tree.js';
 export * from './get-node-env.js';
 export * from './ip-in-networks.js';
 export * from './is-readable-stream.js';
